+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.1.1
 
+- **Add**: `acceptableStatusCodes` property in `DefaultNetworkService`.
+- **Add**: `retry(retryLimit:canRetryClosure:)` to `Observable` extension.
+- **Removed**: `retryWithinErrors` method from `Observable` extension.
 - **Update**: LeadKit to `0.6.x` version
 - **Removed**: `ConnectionError`. (Replaced by `LeadKit.RequestError`)
 - **Removed**: `handleConnectionErrors` from Observable+Extensions